Switch order and detail totals to BigDecimal
Replaced double with BigDecimal for order and order detail totals across backend models, DTOs, and services to improve precision and rounding. Updated calculation logic and method signatures accordingly. Improved payment error handling and UI feedback in PartialPaymentController. Added application icon and fixed payment keypad button layout in payment.fxml.

+++ backendOrderly/src/main/java/com/yebur/backendorderly/order/OrderRequest.java
@@ -1,5 +1,6 @@
 package com.yebur.backendorderly.order;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import jakarta.validation.constraints.NotBlank;
@@ -21,7 +22,7 @@ public class OrderRequest {
     private String paymentMethod;
 
     @Positive
-    private double total;
+    private BigDecimal total;
 
     private Long idEmployee;
